roamwise/pyproject.toml
[project]
name = "roamwise"
version = "0.1.0"
description = "roamwise using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.141.0,<1.0.0",
    "crewai-tools>=0.12.0",
    "panel>=1.3.0",
    "weave>=0.50.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "exa-py>=1.14.16",
]

[project.scripts]
roamwise = "roamwise.main:run"
run_crew = "roamwise.main:run"
train = "roamwise.main:train"
replay = "roamwise.main:replay"
test = "roamwise.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
